EnumeratedType.[] fix.
For some reason, that I remain completely unaware of, the previous
implementation of EnumeratedType.[] would sometimes result in a constant
not being found (i.e. even if e.name did infact == name (proved by puts
staments) the method would not return.

I was not able to reproduce this in any other environment except a
particular project (running under JRuby) and remain completely baffled
by what was causing this. This does fix it and it's perhaps a little
easier to read.